requester/requirements.txt
fastapi==0.95.0
openai==0.27.2
uvicorn==0.21.1
python-dotenv==0.21.0
python-docx==0.8.11
# PyDictionary==2.0.1